requirements.txt
absl-py
altair
amqp
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
astunparse
async-lru
attrs
Babel
backcall
beautifulsoup4
billiard
bleach
blinker
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
comm
coverage
debugpy
decorator
defusedxml
Django
django-cors-headers
django-crispy-forms
django-jazzmin
djangorestframework
djongo
dnspython
docopt
executing
fastjsonschema
filelock
Flask
flatbuffers
fqdn
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
googletrans
grpcio
h11
h2
h5py
hpack
hstspreload
httpcore
httpx
hyperframe
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keras
kombu
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mistune
mpmath
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
oauthlib
opencv-python
opt-einsum
outcome
overrides
p2j
packaging
pandas
pandocfilters
parso
pickleshare
Pillow
pipreqs
platformdirs
prometheus-client
prompt-toolkit
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycparser
pydeck
Pygments
pymongo
Pympler
PySocks
python-dateutil
python-json-logger
pytz
pytz-deprecation-shim
pywin32
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
requests-oauthlib
rfc3339-validator
rfc3986
rfc3986-validator
rich
rpds-py
rsa
scikit-learn
scipy
selenium
Send2Trash
six
smmap
sniffio
sortedcontainers
soupsieve
sqlparse
stack-data
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
terminado
threadpoolctl
tinycss2
toml
toolz
torch
tornado
tqdm
traitlets
trio
trio-websocket
typing_extensions
tzdata
tzlocal
uri-template
urllib3
validators
vine
watchdog
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
wrapt
wsproto
yarg
zipp
